inference reorganization & tests
