+++ sparsecoding/inference/__init__.py
@@ -1,0 +1,21 @@
+from .iht import IHT
+from .inference_method import InferenceMethod
+from .ista import ISTA
+from .lca import LCA
+from .lsm import LSM
+from .mp import MP
+from .omp import OMP
+from .pytorch_optimizer import PyTorchOptimizer
+from .vanilla import Vanilla
+
+__all__ = [
+  'IHT',
+  'InferenceMethod',
+  'ISTA',
+  'LCA',
+  'LSM',
+  'MP',
+  'OMP',
+  'PyTorchOptimizer',
+  'Vanilla'
+]
